[move/#34] JobStageStepStyle -> Style 디렉토리 이동
